Simplify strong overlapping with external function
An additional rayIntersections method has been created which is called
in a semi-static way on two polies. A checking method has been added
which tells if the endpoint was used for the double intersection. This
is because it is still possible for the polies to be drawn within
eachother. This implementation will be fully operational only when
setting the "density" of the polies to 1 will result in an infinite
loop. This current implementation cannot even handle .5 (though it's
correct most of the time). .4 seems to be a good density which achieves
a very low amount of overlap