Add back the modulus checking, removes spotlight
The polies were being generated in a way in which some blocked others if
they were too close. This is because the ray tracing checking was being
done if > 0, which is wrong. It should be % 2 == 1, which signals that
one is inside the other. This make a much more even distribution of
polies and a quicker run time. No anomalies have been noticed yet.